requirements.txt
torch
torchvision
onnx
onnxruntime
opencv-python
pyserial

